MICROBA-921 Move emit_certificate_event() to utils
