+++ lms/djangoapps/certificates/tests/test_api.py
@@ -535,7 +535,7 @@ class GenerateUserCertificatesTest(EventTestMixin, WebCertificateTestMixin, Modu
     ENABLED_SIGNALS = ['course_published']
 
     def setUp(self):  # pylint: disable=arguments-differ
-        super(GenerateUserCertificatesTest, self).setUp('lms.djangoapps.certificates.api.tracker')
+        super().setUp('lms.djangoapps.certificates.utils.tracker')
 
         self.student = UserFactory.create(
             email='joe_user@edx.org',
