+++ lms/djangoapps/certificates/tests/test_webview_views.py
@@ -4,7 +4,6 @@
 
 import datetime
 import json
-from collections import OrderedDict
 from urllib.parse import urlencode
 from uuid import uuid4
 
@@ -26,7 +25,6 @@ from lms.djangoapps.badges.tests.factories import (
     BadgeClassFactory,
     CourseCompleteImageConfigurationFactory
 )
-from lms.djangoapps.certificates.api import get_certificate_url
 from lms.djangoapps.certificates.models import (
     CertificateGenerationCourseSetting,
     CertificateHtmlViewConfiguration,
@@ -36,6 +34,7 @@ from lms.djangoapps.certificates.models import (
     CertificateTemplateAsset,
     GeneratedCertificate
 )
+from lms.djangoapps.certificates.utils import get_certificate_url
 from lms.djangoapps.certificates.tests.factories import (
     CertificateHtmlViewConfigurationFactory,
     GeneratedCertificateFactory,
@@ -1005,7 +1004,7 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
                 response_json = json.loads(response.content.decode('utf-8'))
                 self.assertEqual(CertificateStatuses.generating, response_json['add_status'])
 
-    #TEMPLATES WITHOUT LANGUAGE TESTS
+    # TEMPLATES WITHOUT LANGUAGE TESTS
     @override_settings(FEATURES=FEATURES_WITH_CUSTOM_CERTS_ENABLED)
     @override_settings(LANGUAGE_CODE='fr')
     @patch('lms.djangoapps.certificates.views.webview.get_course_run_details')
@@ -1126,8 +1125,8 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             self.assertContains(response, u'mode: {}'.format(mode))
             self.assertContains(response, 'course name: test_template_1_course')
 
-    ## Templates With Language tests
-    #1
+    # Templates With Language tests
+    # 1
     @override_settings(FEATURES=FEATURES_WITH_CUSTOM_CERTS_ENABLED)
     @override_settings(LANGUAGE_CODE='fr')
     @patch('lms.djangoapps.certificates.views.webview.get_course_run_details')
@@ -1164,16 +1163,16 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             course_id=six.text_type(self.course.id),
             uuid=self.cert.verify_uuid
         )
-        #create a org_mode_and_coursekey template language=null
+        # Create an org_mode_and_coursekey template language=null
         self._create_custom_named_template(
             'test_null_lang_template', org_id=1, mode='honor', course_key=six.text_type(self.course.id), language=None,
         )
-        #verify return template lang = null
+        # Verify return template lang = null
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create a org_mode_and_coursekey template language=wrong_language
+        # Create an org_mode_and_coursekey template language=wrong_language
         self._create_custom_named_template(
             'test_wrong_lang_template',
             org_id=1,
@@ -1181,12 +1180,12 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             course_key=six.text_type(self.course.id),
             language=wrong_language,
         )
-        #verify returns null lang template
+        # Verify returns null lang template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create an org_mode_and_coursekey template language=''
+        # Create an org_mode_and_coursekey template language=''
         self._create_custom_named_template(
             'test_all_languages_template',
             org_id=1,
@@ -1194,12 +1193,12 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             course_key=six.text_type(self.course.id),
             language='',
         )
-        #verify returns null lang template
+        # Verify returns null lang template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_all_languages_template')
 
-        #create a org_mode_and_coursekey template language=lang
+        # Create a org_mode_and_coursekey template language=lang
         self._create_custom_named_template(
             'test_right_lang_template',
             org_id=1,
@@ -1212,7 +1211,7 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_right_lang_template')
 
-    #2
+    # 2
     @override_settings(FEATURES=FEATURES_WITH_CUSTOM_CERTS_ENABLED)
     @patch('lms.djangoapps.certificates.views.webview.get_course_run_details')
     @patch('lms.djangoapps.certificates.api.get_course_organization_id')
@@ -1244,35 +1243,35 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             course_id=six.text_type(self.course.id),
             uuid=self.cert.verify_uuid
         )
-        #create a org and mode template language=null
+        # Create a org and mode template language=null
         self._create_custom_named_template('test_null_lang_template', org_id=1, mode='honor', language=None)
-        #verify return template lang = null
+        # Verify return template lang = null
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create a org and mode template language=wrong_language
+        # Create a org and mode template language=wrong_language
         self._create_custom_named_template('test_wrong_lang_template', org_id=1, mode='honor', language=wrong_language)
-        #verify returns null lang template
+        # Verify returns null lang template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create an org and mode template language=''
+        # Create an org and mode template language=''
         self._create_custom_named_template('test_all_languages_template', org_id=1, mode='honor', language='')
-        #verify returns All Languages template
+        # Verify returns All Languages template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_all_languages_template')
 
-        #create a org and mode template language=lang
+        # Create a org and mode template language=lang
         self._create_custom_named_template('test_right_lang_template', org_id=1, mode='honor', language=right_language)
-        # verify return right_language template
+        # Verify return right_language template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_right_lang_template')
 
-    #3
+    # 3
     @override_settings(FEATURES=FEATURES_WITH_CUSTOM_CERTS_ENABLED)
     @patch('lms.djangoapps.certificates.views.webview.get_course_run_details')
     @patch('lms.djangoapps.certificates.api.get_course_organization_id')
@@ -1302,35 +1301,35 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             course_id=six.text_type(self.course.id),
             uuid=self.cert.verify_uuid
         )
-        #create a org template language=null
+        # Create a org template language=null
         self._create_custom_named_template('test_null_lang_template', org_id=1, language=None)
-        #verify return template lang = null
+        # Verify return template lang = null
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create a org template language=wrong_language
+        # Create a org template language=wrong_language
         self._create_custom_named_template('test_wrong_lang_template', org_id=1, language=wrong_language)
-        #verify returns null lang template
+        # Verify returns null lang template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create an org template language=''
+        # Create an org template language=''
         self._create_custom_named_template('test_all_languages_template', org_id=1, language='')
-        #verify returns All Languages template
+        # Verify returns All Languages template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_all_languages_template')
 
-        #create a org template language=lang
+        # Create a org template language=lang
         self._create_custom_named_template('test_right_lang_template', org_id=1, language=right_language)
-        # verify return right_language template
+        # Verify return right_language template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_right_lang_template')
 
-    #4
+    # 4
     @override_settings(FEATURES=FEATURES_WITH_CUSTOM_CERTS_ENABLED)
     @patch('lms.djangoapps.certificates.views.webview.get_course_run_details')
     @patch('lms.djangoapps.certificates.api.get_course_organization_id')
@@ -1361,30 +1360,30 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             course_id=six.text_type(self.course.id),
             uuid=self.cert.verify_uuid
         )
-        #create a mode template language=null
+        # Create a mode template language=null
         self._create_custom_named_template('test_null_lang_template', mode='honor', language=None)
-        #verify return template with lang = null
+        # Verify return template with lang = null
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create a mode template language=wrong_language
+        # Create a mode template language=wrong_language
         self._create_custom_named_template('test_wrong_lang_template', mode='honor', language=wrong_language)
-        #verify returns null lang template
+        # Verify returns null lang template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create a mode template language=''
+        # Create a mode template language=''
         self._create_custom_named_template('test_all_languages_template', mode='honor', language='')
-        #verify returns All Languages template
+        # Verify returns All Languages template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_all_languages_template')
 
-        #create a mode template language=lang
+        # Create a mode template language=lang
         self._create_custom_named_template('test_right_lang_template', mode='honor', language=right_language)
-        # verify return right_language template
+        # Verify return right_language template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_right_lang_template')
@@ -1423,28 +1422,28 @@ class CertificatesViewsTests(CommonCertificatesTestCase, CacheIsolationTestCase)
             course_id=six.text_type(self.course.id),
             uuid=self.cert.verify_uuid
         )
-        #create a mode template language=null
+        # Create a mode template language=null
         self._create_custom_named_template('test_null_lang_template', org_id=1, mode='honor', language=None)
-        #verify return template with lang = null
+        # Verify return template with lang = null
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create a mode template language=wrong_language
+        # Create a mode template language=wrong_language
         self._create_custom_named_template('test_wrong_lang_template', org_id=1, mode='honor', language=wrong_language)
-        #verify returns null lang template
+        # Verify returns null lang template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_null_lang_template')
 
-        #create a mode template language=''
+        # Create a mode template language=''
         self._create_custom_named_template('test_all_languages_template', org_id=1, mode='honor', language='')
-        #verify returns All Languages template
+        # Verify returns All Languages template
         response = self.client.get(test_url)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, 'course name: test_all_languages_template')
 
-        #create a mode template language=lang
+        # Create a mode template language=lang
         self._create_custom_named_template('test_right_lang_template', org_id=1, mode='honor', language=right_language)
         # verify return right_language template
         response = self.client.get(test_url)
